Separate agent skill policy enforcement
The agent runtime loop still carried skill-policy parsing, enforcement, and active-skill hydration inline. Moving that behavior into a focused runtime module keeps the loop thinner while preserving the existing runtime index exports for policy helpers.

Constraint: Public runtime helper exports must remain stable during decomposition.
Rejected: Add a new role filter during hydration | the old inline helper scanned all messages, so this slice preserves that behavior exactly.
Confidence: high
Scope-risk: narrow
Directive: Do not change skill hydration semantics while splitting runtime code unless a dedicated behavior-change test and migration note are added.
Tested: deno test --no-check --allow-all src/agent/runtime/skill-policy.test.ts
Tested: deno test --no-check --allow-all src/agent/runtime/skill-policy.test.ts src/agent/runtime/refresh.test.ts src/agent/runtime/load-skill-tool.test.ts
Tested: deno test --no-check --allow-all src/agent/runtime
Tested: deno fmt --check src/agent/runtime/index.ts src/agent/runtime/skill-policy-enforcement.ts src/agent/runtime/skill-policy.test.ts
Tested: deno lint src/agent/runtime/index.ts src/agent/runtime/skill-policy-enforcement.ts src/agent/runtime/skill-policy.test.ts
Tested: deno check --frozen src/agent/runtime/index.ts src/agent/runtime/skill-policy-enforcement.ts src/agent/runtime/skill-policy.test.ts
Tested: deno task verify:quick
Not-tested: Hosted CI binary e2e before PR creation
